+++ tests/python/sdk/util.py
@@ -9,11 +9,11 @@ from urllib.parse import urlparse
 
 import pytest
 from cvat_sdk.api_client.rest import RESTClientObject
-from cvat_sdk.core.helpers import TqdmProgressReporter2
+from cvat_sdk.core.helpers import DeferredTqdmProgressReporter
 
 
 def make_pbar(file, **kwargs):
-    return TqdmProgressReporter2({"file": file, "mininterval": 0, **kwargs})
+    return DeferredTqdmProgressReporter({"file": file, "mininterval": 0, **kwargs})
 
 
 def generate_coco_json(filename: Path, img_info: Tuple[Path, int, int]):